atoi, lstclear, strlcat were changed the rest is the same

+++ Core/libft/ft_atoi.c
@@ -1,10 +1,22 @@
-#include "libft.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: magulyas <magulyas@student.42belgium.be    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/04/09 17:13:15 by magulyas          #+#    #+#             */
+/*   Updated: 2026/05/12 17:33:12 by magulyas         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+// Converts the initial portion of the string
 
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
